% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_routines.R
\name{adj_list_2_matrix}
\alias{adj_list_2_matrix}
\title{Convert adjacency list to a matrix}
\usage{
adj_list_2_matrix(adj_list)
}
\arguments{
\item{adj_list}{The adjacency list of the tree.}
}
\value{
A symmetric matrix of 1's and 0's, with 1 in entry (i,j) representing
an edge between the two vertices.
}
\description{
Converts a list that represents a tree into a binary matrix.
}
\details{
Remember that the list has to be for a tree, not a general graph. Please
see other help pages for more specifications.

This is a low-level function, used within the S4 class TreeHarp. It is not
generally meant for use by the user.

It works by filling up the upper diagonal of the matrix before reflecting it.
}
