% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestharp.R
\name{apply_can_improve_this}
\alias{apply_can_improve_this}
\title{Identifies if an apply function can improve the code}
\usage{
apply_can_improve_this(fname, window_size = 2)
}
\arguments{
\item{fname}{A filename: either Rmd/qmd or R.}

\item{window_size}{A window size to analyse.}
}
\value{
A numeric vector corresponding to windows for review is returned. If 
none are found, then NULL is returned.
}
\description{
Identifies if an apply function can improve the code
}
\details{
First, all library() calls and read() calls are removed. Next,
\code{only_actual_args_differ} is run on all pairwise expressions. The 
matrix is searched for blocks of 1's along the diagonal.
}
