% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.R
\name{as.matrix}
\alias{as.matrix}
\title{TreeHarp Cast a TreeHarp to Matrix.}
\arguments{
\item{from}{A treeharp object.}
}
\value{
A matrix.
}
\description{
Convert a treeharp object to an adjacency matrix.
}
