% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoharp.R
\docType{package}
\name{autoharp}
\alias{autoharp}
\title{autoharp: Semi-Automatic Grading of R and Rmd Scripts}
\description{
The autoharp package provides functions for running and analysing 
R script/Rmd submissions from students.
}
\section{Main functions are}{

\enumerate{
  \item \code{\link{populate_soln_env}}
  \item \code{\link{render_one}}
  \item \code{\link{run_tuner}}
  \item \code{\link{TreeHarp}}
}

The user manuals can be found at \url{https://singator.github.io/autoharp-docs/}
}

