% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_runtime.R
\name{check_runtime}
\alias{check_runtime}
\title{Calculate Run-time Statistics}
\usage{
check_runtime(stud_fname, knit_root_dir, return_env = FALSE)
}
\arguments{
\item{stud_fname}{The rmd filename of the student.}

\item{knit_root_dir}{The working directory to use when knitting the file.}

\item{return_env}{A logical value to indicate if the environment from the
rmd file should be return. If FALSE, an NA value is returned.}
}
\value{
A list containing the running time in seconds, the memory used
by the final environment in bytes (as a numeric scalar), and the
environment object containing all the generated objects from the rmd file.
}
\description{
This is stand-alone function. It computes the runtime stats
without rendering the md/html/pdf file.
}
\details{
This routine is not used within any other function within the 
package. Figures are not cleaned or removed.
}
\seealso{
\code{\link{render_one}}
}
