% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lints.R
\name{count_lints_all}
\alias{count_lints_all}
\title{Lint counter}
\usage{
count_lints_all(file_names, lint_list, lint_labels)
}
\arguments{
\item{file_names}{The path to the rmd files that need to be checked for
lints.}

\item{lint_list}{List of lints to check for.}

\item{lint_labels}{List of labels to name the vector to return.}
}
\value{
Dataframe containing the lints.
}
\description{
Count number of lints in one folder
}
\details{
The function will count the number of lints in a file. The lints to be
checked can be passed as an argument. Else, the default will be used.
The defaults are as follows:
\itemize{
\item T_and_F_symbol_linter
\item line_length_linter
\item assignment_linter
\item absolute_path_linter
\item pipe_continuation_linter
}

Note that labels would also need to be given if the non-default lints
are chosen.
}
