% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lints.R
\name{count_lints_one}
\alias{count_lints_one}
\title{File lint counter}
\usage{
count_lints_one(rmd_file, lint_list, lint_labels)
}
\arguments{
\item{rmd_file}{The path to the rmd file to check for lints.}

\item{lint_list}{List of lints to check for.}

\item{lint_labels}{List of labels to name the vector to return.}
}
\value{
Vector containing the lints.
}
\description{
Count number of lints in one file
}
\details{
The function will count the number of lints in a file. The lints to be 
checked can be passed as an argument. Else, the default will be used. 
The defaults are as follows: 
* T_and_F_symbol_linter
* line_length_linter
* assignment_linter
* absolute_path_linter
* pipe_continuation_linter
Note that labels would also need to be given if the non-default lints 
are chosen.
}
