% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplify_to_r.R
\name{examplify_to_r}
\alias{examplify_to_r}
\title{Convert from Examplify to R}
\usage{
examplify_to_r(in_fname, out_fname, verbose = FALSE)
}
\arguments{
\item{in_fname}{A file name of a student submission html file.}

\item{out_fname}{The output R script.}

\item{verbose}{Controls verbosity of output.}
}
\value{
This will return NULL, but will generate an R script as output.
}
\description{
This function converts an Examplify script (from a student,html)
into an R script.
}
\details{
The script has to be exported in html format, using a particular profile.
The questions are stored in tags nested under id = "answers" and tag type "h2".
The answers are stored in tags nested under id = "answers" and class
"content".

Some of the student text may contain R code and text mixed up, so
tidy_source() may not work on those, since it is parsed.
}
