% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_non_chunks}
\alias{extract_non_chunks}
\title{Extract non-chunks from an Rmd file.}
\usage{
extract_non_chunks(rmd_name, out_name)
}
\arguments{
\item{rmd_name}{A character string, the name of the rmd file to get the
chunks from.}

\item{out_name}{An output filename, to dump the text to.}
}
\value{
If out_name is missing, then a character vector is returned. If 
outfname is specified, then nothing is returned. The text is written to the 
file instead.
}
\description{
Extracts non-chunks from an Rmd file.
}
