% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracton_section_text.R
\name{extract_section_text}
\alias{extract_section_text}
\title{Extract section text from Rmd}
\usage{
extract_section_text(rmd_name, hdr_pat, ignore_case = TRUE)
}
\arguments{
\item{rmd_name}{The filename of the Rmd script.}

\item{hdr_pat}{The regular expression pattern to pick up the section title.}

\item{ignore_case}{A boolean - whether or not to ignore case when matching for
the section title.}
}
\value{
Returns a character vector containing all the text written in the section
that begins with the specified pattern.

The pattern should pick up a unique section/sub-section/sub-sub-section.
Otherwise, it will stop and raise an error.
}
\description{
Extract section text from Rmd
}
\details{
The text that is picked up begins with the specified section, and ends with
the next string of pound symbols (#)
}
