% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_getter-length.R
\name{get_adj_list}
\alias{get_adj_list}
\alias{get_adj_list,TreeHarp-method}
\title{Generic for Getting Adjacency List}
\usage{
get_adj_list(x, ...)

\S4method{get_adj_list}{TreeHarp}(x, ...)
}
\arguments{
\item{x}{An object of class TreeHarp.}

\item{...}{Unused arguments, for now.}
}
\value{
The adjacency list for a TreeHarp object.
}
\description{
The generic method definition for getting adjacency list from a
TreeHarp object.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_adj_list(TreeHarp)}: A getter.

Allows user to extract the adjacency list of a treeharp object.

}}
