% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_getter-length.R
\name{get_child_ids}
\alias{get_child_ids}
\alias{get_child_ids,TreeHarp-method}
\alias{get_child_ids,list-method}
\title{Generic for Getting Child Node Ids}
\usage{
get_child_ids(x, node_num)

\S4method{get_child_ids}{TreeHarp}(x, node_num)

\S4method{get_child_ids}{list}(x, node_num)
}
\arguments{
\item{x}{An object of class TreeHarp.}

\item{node_num}{An integer, length 1. This the node whose children we are
after. If the specified node is a leaf, the NULL is returned.}
}
\value{
An integer vector, indicating the children node ids.
}
\description{
The generic method definition for getting child node ids.
}
\section{Methods (by class)}{
\itemize{
\item \code{TreeHarp}: Obtain child nodes.

Allows user to extract the child nodes from a specified node from TreeHarp
object.

\item \code{list}: Obtain child nodes.

Allows user to extract the child nodes from a specified node from an 
adjacency list.
}}

\seealso{
\code{\link{get_parent_id}}
}
