% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_routines.R
\name{get_child_ids2}
\alias{get_child_ids2}
\title{Get the children node ids}
\usage{
get_child_ids2(adj_list, at_node)
}
\arguments{
\item{adj_list}{The adjacency list of the tree.}

\item{at_node}{The node whose children should be extracted.}
}
\value{
A vector of integers specifying the children of that particular node.
If the node is a leaf, it returns NULL.
}
\description{
This function retrieves the child node ids of a given node from an
adjacency list of a tree.
}
\details{
Remember that the list has to be for a tree, not a general graph. Please
see other help pages for more specifications.

This is a low-level function, used within the S4 class TreeHarp. It is not
generally meant for use by the user.
}
