% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_2_tree_helpers.R
\name{get_next_depth_id}
\alias{get_next_depth_id}
\title{Get the id and depth of a child node.}
\usage{
get_next_depth_id(parent_node_id, env_ni)
}
\arguments{
\item{parent_node_id}{The id of the parent node we are considering.}

\item{env_ni}{An environment object, possibly containing a data frame with
columns id, name, call_status, arg_type and depth.}
}
\value{
A list containing the id and depth of the next node.
}
\description{
From the parent's depth and the last labelled node, we obtain the node id 
and depth of a child.
}
\details{
This is for internal use. It may be removed from user-view soon!
}
