% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_getter-length.R
\name{get_node_types}
\alias{get_node_types}
\alias{get_node_types,TreeHarp-method}
\title{Generic for Getting Node Types}
\usage{
get_node_types(x, ...)

\S4method{get_node_types}{TreeHarp}(x, ...)
}
\arguments{
\item{x}{An object of class TreeHarp.}

\item{...}{Unused arguments, for now.}
}
\value{
A data frame containing the node types for a TreeHarp object. If the 
slot is empty, NA is returned.
}
\description{
The generic method definition for getting node types from a
TreeHarp object.
}
\section{Methods (by class)}{
\itemize{
\item \code{TreeHarp}: A getter.

Allows user to extract the node types of a treeharp object.
}}

