% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tuner.R
\name{get_summary_output}
\alias{get_summary_output}
\title{Function to extract the summary content}
\usage{
get_summary_output(
  rmd_file,
  summary_header = "# Summary Output",
  dir = tempdir()
)
}
\arguments{
\item{rmd_file}{The path to the rmd file to search for the summary.}

\item{summary_header}{The header to look for.}

\item{dir}{A temporary directory to store the temporary Rmarkdown file
before extracting the html content. The temp file will be deleted
before the function exits.}
}
\value{
The function is used as a helper function. Returns the HTML 
formatted string.
}
\description{
This function will look for the explanation for the checks being done. 
If there is an explanation, the function will return the summary in HTML 
format. If not it will return 'not found' in HTML format.
}
