% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_treeharp.R
\name{join_treeharps}
\alias{join_treeharps}
\title{Root a list of trees.}
\usage{
join_treeharps(...)
}
\arguments{
\item{...}{A list of Treeharp objects.}
}
\value{
A TreeHarp object
}
\description{
Given a list of trees, this will root them.
}
\details{
This function combines TreeHarp objects into a single TreeHarp.
The function will root all of them at a node called "script", which is 
neither a function call nor an argument nor a symbol. The BFS ordering is 
then updated.

Objects that are not of class TreeHarp will be dropped from the list 
before the rooting takes place.
}
