% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestharp.R
\name{pipe_can_improve_this}
\alias{pipe_can_improve_this}
\title{Identify if use of pipe operator can improve a code section}
\usage{
pipe_can_improve_this(fname, window_len = 2)
}
\arguments{
\item{fname}{A filename - either a Rmd/qmd or R script.}

\item{window_len}{A window length to analyse.}
}
\value{
A numeric vector corresponding to start of the window to be reviewed.
If no lines are found, NULL is returned.
}
\description{
Identify if use of pipe operator can improve a code section
}
\details{
A rolling window approach is used here. For each 
expression, the assigned object is extracted (if any). Subsequent lines are 
checked to see if this object appears as an actual argument.

Window length two means that only one more line is checked.
}
