% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/th_getter-length.R
\name{plot,TreeHarp-method}
\alias{plot,TreeHarp-method}
\title{TreeHarp Plotting TreeHarp Objects}
\usage{
\S4method{plot}{TreeHarp}(x, y, ...)
}
\arguments{
\item{x}{An object of class TreeHarp.}

\item{y}{Unused.}

\item{...}{Additional arguments passed on to plot.igraph().}
}
\value{
Returns NULL, invisibly.
}
\description{
A plot method for visualising treeharp objects.
}
\details{
The treeharp object is converted to an igraph object before it
is plotted.
}
