% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_2_tree_helpers.R
\name{rbind_to_nodes_info}
\alias{rbind_to_nodes_info}
\title{Update node information.}
\usage{
rbind_to_nodes_info(id, name, call_status, formal_arg, depth, env_ni)
}
\arguments{
\item{id}{The id of the node to be added. This should be an integer
of length 1.}

\item{name}{The name of the node.}

\item{call_status}{Is the language object a call or a symbol/literal? This 
should a logical value.}

\item{formal_arg}{Is the language object a formal argument or not? This 
should be a logical value.}

\item{depth}{An integer indicating the depth of this language object in 
the parse tree.}

\item{env_ni}{An environment object, possibly containing a data frame with
columns id, name, call_status, formal_arg and depth.}
}
\value{
TRUE is returned invisibly.
}
\description{
Updates the node information regarding an R expression.
}
\details{
This is for internal use. It may be removed from user-view soon!
}
