% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_related.R
\name{rmd_to_token_count}
\alias{rmd_to_token_count}
\title{Count tokens in R/Rmd}
\usage{
rmd_to_token_count(fname, include_actuals = TRUE)
}
\arguments{
\item{fname}{The Rmd or R file name.}

\item{include_actuals}{Whether actual arguments/literals should be included.
If this is FALSE, then only calls and formal arguments will be used in
the count.}
}
\value{
A tibble. The tibble will contain a the frequency count for all 
tokens present in the student script.
}
\description{
Count the individual tokens. Part of the NLP analysis process.
}
