% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_BFS.R
\name{to_BFS}
\alias{to_BFS}
\title{Function to rearrage nodes in BFS}
\usage{
to_BFS(adj_list, node_info)
}
\arguments{
\item{adj_list}{The output of lang_2_tree.}

\item{node_info}{The output of lang_2_tree.}
}
\value{
An adjacency list and nodes info data frame in BFS order.
}
\description{
Function to rearrage nodes in BFS
}
\details{
This function is for an internal TreeHarp constructor use. It is
not exported.
}
