% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_kernel.R
\name{tree_sim}
\alias{tree_sim}
\title{Compute tree similarity}
\usage{
tree_sim(t1, t2, norm = FALSE, ...)
}
\arguments{
\item{t1}{A TreeHarp object}

\item{t2}{Anothe TreeHarp object.}

\item{norm}{A logical value, indicating if the kernel function should be
normalised, to account for different tree lengths.}

\item{...}{Unused arguments, reserved for mcmapply}
}
\value{
A numerical value between 0 and 1 (if normed).
}
\description{
Computes similarity between two trees (non-recursively)
}
