% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang_2_tree_helpers.R
\name{update_adj_list}
\alias{update_adj_list}
\title{Update adjacency list.}
\usage{
update_adj_list(
  update_type = c("new_node", "add_child"),
  node_id,
  node_name,
  child_node,
  env_ni
)
}
\arguments{
\item{update_type}{This should be either "new_node" or "add_child". If it 
is a new node, an empty list component is added. If it is add child, 
then child_node should be provided too.}

\item{node_id}{An integer.}

\item{node_name}{The name of the new node to be added. This must be provided 
if the update_type is "new_node".}

\item{child_node}{An integer.}

\item{env_ni}{An environment object, possibly containing an adjacency list 
that will later be used to construct a TreeHarp object.}
}
\value{
An invisible TRUE is returned.
}
\description{
Updates the adjacency list for an R expression parse tree.
}
\details{
This is for internal use. It may be removed from user-view soon!
}
